Remove modulo operation
Temporarily removed support for the modulo (%) operation since it adds unneccessary complexity. Modulo is only defined for integer operands, while the calculator supports operands of type double. This means that validation has to be added to check if the user has entered numbers that can be safely converted to integers (e.g. 5.0 to 5). This is something that is best done after a working calculator has been created.